pyproject.toml
[project]
name = "langchain_rag"
version = "0.1.0"
description = "RAG pipeline with LangChain, FastAPI and Uvicorn"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "qdrant-client>=1.17.0",
    "torch>=2.10.0",
    "transformers>=5.2.0",
    "uvicorn>=0.41.0",
    "python-dotenv>=1.2.2",
    "langchain-qdrant>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.10",
    "pypdf>=6.7.5",
    "sentence-transformers>=5.2.3"
]
